% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_h5ad.R
\name{download_source_h5ad}
\alias{download_source_h5ad}
\title{Download source H5AD to local file name.}
\usage{
download_source_h5ad(
  dataset_id,
  file,
  overwrite = FALSE,
  census_version = "stable",
  census = NULL
)
}
\arguments{
\item{dataset_id}{The dataset_id of interest.}

\item{file}{Local file name to store H5AD file.}

\item{overwrite}{TRUE to allow overwriting an existing file.}

\item{census_version}{The desired Census version.}

\item{census}{An open Census handle for \code{census_version}. If not provided, then
it will be opened and closed automatically; but it's more efficient
to reuse a handle if calling \code{download_source_h5ad()} multiple times.}
}
\description{
Download source H5AD to local file name.
}
\examples{
download_source_h5ad("0895c838-e550-48a3-a777-dbcd35d30272", "/tmp/data.h5ad", overwrite = TRUE)
}
