% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_directory.R
\name{get_census_mirror}
\alias{get_census_mirror}
\title{Get locator information about a Census mirror}
\usage{
get_census_mirror(mirror)
}
\arguments{
\item{mirror}{Name of the mirror.}
}
\value{
List with mirror information
}
\description{
Get locator information about a Census mirror
}
\examples{
get_census_mirror("AWS-S3-us-west-2")
}
