% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_directory.R
\name{get_census_mirror_directory}
\alias{get_census_mirror_directory}
\title{Get the directory of Census mirrors currently available}
\usage{
get_census_mirror_directory()
}
\value{
Nested list with information about available mirrors
}
\description{
Get the directory of Census mirrors currently available
}
\examples{
get_census_mirror_directory()
}
