% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_directory.R
\name{get_census_version_description}
\alias{get_census_version_description}
\title{Get release description for a Census version}
\usage{
get_census_version_description(census_version)
}
\arguments{
\item{census_version}{The census version name.}
}
\value{
List with the release location and metadata
}
\description{
Get release description for a Census version
}
\examples{
as.data.frame(get_census_version_description("stable"))
}
