% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_directory.R
\name{get_census_version_directory}
\alias{get_census_version_directory}
\title{Get the directory of Census releases currently available}
\usage{
get_census_version_directory()
}
\value{
Data frame of available cell census releases, including location and
metadata.
}
\description{
Get the directory of Census releases currently available
}
\examples{
get_census_version_directory()
}
