% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_helpers.R
\name{get_presence_matrix}
\alias{get_presence_matrix}
\title{Read the feature dataset presence matrix.}
\usage{
get_presence_matrix(census, organism, measurement_name = "RNA")
}
\arguments{
\item{census}{The census object from \code{cellxgene.census::open_soma()}.}

\item{organism}{The organism to query, usually one of \verb{Homo sapiens} or \verb{Mus musculus}}

\item{measurement_name}{The measurement object to query. Defaults to \code{RNA}.}
}
\value{
A \code{tiledbsoma::matrixZeroBasedView} object with dataset join id & feature
join id dimensions, filled with 1s indicating presence. The sparse matrix
is accessed with zero-based indexes since the join id's may be zero.
}
\description{
Read the feature dataset presence matrix.
}
\examples{
census <- open_soma()
on.exit(census$close(), add = TRUE)
print(get_presence_matrix(census, "Homo sapiens")$dim())
}
