% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_soma}
\alias{open_soma}
\title{Open the Census}
\usage{
open_soma(
  census_version = "stable",
  uri = NULL,
  tiledbsoma_ctx = NULL,
  mirror = NULL
)
}
\arguments{
\item{census_version}{The version of the Census, e.g., "stable".}

\item{uri}{A URI containing the Census SOMA objects to open instead of a
released version. (If supplied, takes precedence over
\code{census_version}.)}

\item{tiledbsoma_ctx}{A \code{tiledbsoma::SOMATileDBContext} built using
\code{new_SOMATileDBContext_for_census()}. Optional (created automatically)
if using \code{census_version} and the context does not need to be reused.}

\item{mirror}{The Census mirror to access; one of \code{names(get_census_mirror_directory())}.}
}
\value{
Top-level \code{tiledbsoma::SOMACollection} object. After use, the census
should be closed to release memory and other resources, usually with
\code{on.exit(census$close(), add = TRUE)}. Closing the top-level census
will also close all SOMA objects accessed through it.
}
\description{
Open the Census
}
\examples{
census <- open_soma()
as.data.frame(census$get("census_info")$get("summary")$read()$concat())
census$close()
}
