% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_helpers.R
\name{get_seurat}
\alias{get_seurat}
\title{Convenience wrapper around \code{SOMAExperimentAxisQuery}, to build and execute a
query, and return it as a \code{Seurat} object.}
\usage{
get_seurat(
  census,
  organism,
  measurement_name = "RNA",
  X_name = "raw",
  obs_query = NULL,
  obs_column_names = NULL,
  var_query = NULL,
  var_column_names = NULL
)
}
\arguments{
\item{census}{The census object, usually returned by \code{cellxgene.census::open_soma()}.}

\item{organism}{The organism to query, usually one of \verb{Homo sapiens} or \verb{Mus musculus}}

\item{measurement_name}{The measurement object to query. Defaults to \code{RNA}.}

\item{X_name}{The \code{X} layer to query. Defaults to \code{raw}.}

\item{obs_query}{A \code{SOMAAxisQuery} for the \code{obs} axis.}

\item{obs_column_names}{Columns to fetch for the \code{obs} data frame.}

\item{var_query}{A \code{SOMAAxisQuery} for the \code{var} axis.}

\item{var_column_names}{Columns to fetch for the \code{var} data frame.}
}
\value{
A \code{Seurat} object containing the sensus slice.
}
\description{
Convenience wrapper around \code{SOMAExperimentAxisQuery}, to build and execute a
query, and return it as a \code{Seurat} object.
}
