% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{new_SOMATileDBContext_for_census}
\alias{new_SOMATileDBContext_for_census}
\title{Create \code{SOMATileDBContext} for Census}
\usage{
new_SOMATileDBContext_for_census(
  census_version_description,
  mirror = "default",
  ...
)
}
\arguments{
\item{census_version_description}{The result of \code{get_census_version_description()}
for the desired Census version.}

\item{mirror}{The name of the intended census mirror (or \code{get_census_mirror_directory()[[name]]}
to save the lookup), or NULL to configure for local file access.}

\item{...}{Custom configuration options.}
}
\value{
SOMATileDBContext object for \code{open_soma()}.
}
\description{
Create a SOMATileDBContext suitable for using with \code{open_soma()}.
Typically \code{open_soma()} creates a context automatically, but one can be created
separately in order to set custom configuration options, or to share it between
multiple open Census handles.
}
\examples{
census_desc <- get_census_version_description("stable")
ctx <- new_SOMATileDBContext_for_census(census_desc, "soma.init_buffer_bytes" = paste(4 * 1024**3))
census <- open_soma("stable", tiledbsoma_ctx = ctx)
census$close()
}
